+++ src/main/java/com/junwoo/ott/global/search/dto/body/VideoRandomDto.java
@@ -1,15 +1,16 @@
 package com.junwoo.ott.global.search.dto.body;
 
+import com.junwoo.ott.global.search.dto.response.VideoResponseDto;
 import lombok.Getter;
 
 @Getter
-public class VideoRandomSearchDto {
+public class VideoRandomDto {
 
   private final Long videoId;
   private final String title;
   private final String posterUrl;
 
-  public VideoRandomSearchDto(VideoSearchDto videoSearchDto) {
+  public VideoRandomDto(VideoResponseDto videoSearchDto) {
     this.videoId = videoSearchDto.getVideo_id();
     this.title = videoSearchDto.getTitle();
     this.posterUrl = videoSearchDto.getPoster_url();
